unify input type aliases
